fix(#30): list null 오류 수정
