+++ src/main/java/com/neo/neomarket/dto/Auction/request/AuctionPostUpdateDTO.java
@@ -1,4 +1,4 @@
-package com.neo.neomarket.dto.request;
+package com.neo.neomarket.dto.Auction.request;
 
 
 import lombok.*;
@@ -10,14 +10,11 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Setter
-public class AuctionPostCreateDTO {
+public class AuctionPostUpdateDTO {
     private String title;
     private List<String> pictureUrls;
     private String content;
-    private Long startPrice;
-    private Long currentPrice;
     private String deadline;
     private String category;
-    private Long userId;
 
 }
