+++ src/main/java/com/neo/neomarket/dto/Auction/request/response/AuctionPostDTO.java
@@ -1,4 +1,4 @@
-package com.neo.neomarket.dto.response;
+package com.neo.neomarket.dto.Auction.request.response;
 
 
 import lombok.*;
@@ -10,15 +10,17 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Setter
-public class AuctionPostReadDTO {
+public class AuctionPostDTO {
     private Long id;
     private String title;
+    //private String pictureUrl;
     private List<String> pictureUrls;
     private String content;
-    private String deadline;
-    private String category;
     private Long startPrice;
     private Long currentPrice;
-    private String nickname;
+    private String deadline;
+    private String category;
+    private int status;
+    private Long userId;
 
 }
