+++ src/main/java/com/neo/neomarket/dto/Auction/request/response/AuctionPostReadDTO.java
@@ -1,4 +1,4 @@
-package com.neo.neomarket.dto.response;
+package com.neo.neomarket.dto.Auction.request.response;
 
 
 import lombok.*;
@@ -10,17 +10,15 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Setter
-public class AuctionPostDTO {
+public class AuctionPostReadDTO {
     private Long id;
     private String title;
-    //private String pictureUrl;
     private List<String> pictureUrls;
     private String content;
-    private Long startPrice;
-    private Long currentPrice;
     private String deadline;
     private String category;
-    private int status;
-    private Long userId;
+    private Long startPrice;
+    private Long currentPrice;
+    private String nickname;
 
 }
