+++ src/main/java/com/neo/neomarket/service/AuctionService.java
@@ -1,10 +1,10 @@
 package com.neo.neomarket.service;
 
 import com.neo.neomarket.dto.*;
-import com.neo.neomarket.dto.request.AuctionPostCreateDTO;
-import com.neo.neomarket.dto.request.AuctionPostUpdateDTO;
-import com.neo.neomarket.dto.response.AuctionPostDTO;
-import com.neo.neomarket.dto.response.AuctionPostReadDTO;
+import com.neo.neomarket.dto.Auction.request.AuctionPostCreateDTO;
+import com.neo.neomarket.dto.Auction.request.AuctionPostUpdateDTO;
+import com.neo.neomarket.dto.Auction.request.response.AuctionPostDTO;
+import com.neo.neomarket.dto.Auction.request.response.AuctionPostReadDTO;
 
 import java.util.List;
